Created a new Business Service Layer for Submissions
Created a new service in between the Submissions controller and de Data Access Service, so that it makes more sense to extract Business Logic methods.

Currently the only additional methods needed are two methods to parse the Submission's tags String, but this will make it easier to unit test and improves separation of concerns.